Decouple marker management and XML parsing from PackStateService
