+++ KXMapStudio.App/Services/Pack/DirectoryPackLoader.cs
@@ -6,6 +6,15 @@ namespace KXMapStudio.App.Services.Pack;
 
 public class DirectoryPackLoader : IPackLoader
 {
+    private readonly MarkerXmlParser _markerXmlParser;
+    private readonly CategoryBuilder _categoryBuilder;
+
+    public DirectoryPackLoader(MarkerXmlParser markerXmlParser, CategoryBuilder categoryBuilder)
+    {
+        _markerXmlParser = markerXmlParser;
+        _categoryBuilder = categoryBuilder;
+    }
+
     public async Task<PackLoadResult> LoadPackAsync(string path)
     {
         var originalRawContent = new Dictionary<string, byte[]>(StringComparer.OrdinalIgnoreCase);
@@ -16,7 +25,7 @@ public class DirectoryPackLoader : IPackLoader
             originalRawContent[relativePath] = await File.ReadAllBytesAsync(file);
         }
 
-        var packLoader = new PackLoader();
+        var packLoader = new PackLoader(_markerXmlParser, _categoryBuilder);
         return await packLoader.LoadPackFromMemoryAsync(originalRawContent, path, false);
     }
 }
