+++ KXMapStudio.App/Services/Pack/PackLoaderFactory.cs
@@ -4,6 +4,15 @@ namespace KXMapStudio.App.Services.Pack;
 
 public class PackLoaderFactory
 {
+    private readonly MarkerXmlParser _markerXmlParser;
+    private readonly CategoryBuilder _categoryBuilder;
+
+    public PackLoaderFactory(MarkerXmlParser markerXmlParser, CategoryBuilder categoryBuilder)
+    {
+        _markerXmlParser = markerXmlParser;
+        _categoryBuilder = categoryBuilder;
+    }
+
     public IPackLoader GetLoader(string path)
     {
         if (File.Exists(path))
@@ -11,16 +20,16 @@ public class PackLoaderFactory
             var extension = Path.GetExtension(path).ToLowerInvariant();
             if (extension is ".taco" or ".zip")
             {
-                return new ArchivePackLoader();
+                return new ArchivePackLoader(_markerXmlParser, _categoryBuilder);
             }
             else if (extension == ".xml")
             {
-                return new SingleFilePackLoader();
+                return new SingleFilePackLoader(_markerXmlParser, _categoryBuilder);
             }
         }
         else if (Directory.Exists(path))
         {
-            return new DirectoryPackLoader();
+            return new DirectoryPackLoader(_markerXmlParser, _categoryBuilder);
         }
 
         throw new FileNotFoundException("The specified pack, file, or directory was not found.", path);
